pyproject.toml
[project]
name = "Fontweaver"
version = "0.1.0"
description = "AI training code"
authors = [{ name = "Chris Lai", email = "lcmchristopher@gmail.com" }]
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.0,<4",
    "numpy>=2.1.2,<3",
    "lightning>=2.5.0.post0,<3",
    "openai>=1.76.0,<2",
    "freetype-py>=2.5.1,<3",
    "gradio>=5.16.0,<6",
    "pillow>=11.0.0,<12",
    "ipykernel>=6.29.5,<7",
    "mypy>=1.15.0,<2",
    "tensorboard>=2.19.0,<3",
    "tensorboardx>=2.6.2.2,<3",
    "x-transformers>=2.1.37,<3",
    "fonttools>=4.56.0,<5",
    "torch>=2.6.0,<3",
    "transformers>=4.48.3,<5",
    "supabase>=2.15.0,<3",
    "pytest>=8.3.5",
    "dotenv>=0.9.9",
    "typer>=0.15.1",
    "google-genai>=1.13.0",
    "bitsandbytes>=0.45.5",
    "lightning-sdk>=0.2.15",
    "pyarrow>=21.0.0",
    "lxml>=6.0.0",
]

[project.scripts]
fontweaver = "fontweaver:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fontweaver"]
